Updated Show more button
