+++ src/components/about/AboutPage.styled.tsx
@@ -47,8 +47,6 @@ export const LinkedInText = styled(Typography)(() => ({
 
 export const Description = styled(Typography)(() => ({
   textAlign: 'initial',
-  marginBottom: theme.spacing(3),
-  overflow: 'hidden',
 }))
 
 export const AboutWrapper = styled(Grid)(() => ({
